changed int -> float in the current sensor definitions

+++ motor_stand_slave/src/motor_stand_slave.cpp
@@ -49,8 +49,8 @@ bool see_object;
 
 const int CURRENT_PIN = A2;
 const int VOLTAGE_PIN = A3;
-const int CURRENT_SENSITIVITY = 0.020;
-int ZERO_CURRENT_VOLTAGE = 0.22;
+const float CURRENT_SENSITIVITY = 0.020;
+float ZERO_CURRENT_VOLTAGE = 0.22;
 
 ///////////////////////////////////////////////////////////////////////////////////////
 // TIMING VARIABLE DEFINITIONS (FOR TRACKING)
